+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for pandoc-link-context
 
+## Unreleased
+
+- Include Link attributes (excluded id/class) in results 
+
 ## 1.0.0.0
 
 * First version. Released on an unsuspecting world.